Refactor all Object related requests to one file.
This is part of the work for #838. It consolidate the request types in
one file, like we have done for Bucket and the *Acl things.
