+++ google/cloud/storage/internal/object_requests_test.cc
@@ -1,0 +1,257 @@
+// Copyright 2018 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "google/cloud/storage/internal/object_requests.h"
+#include <gmock/gmock.h>
+
+namespace google {
+namespace cloud {
+namespace storage {
+inline namespace STORAGE_CLIENT_NS {
+namespace internal {
+namespace {
+using ::testing::HasSubstr;
+
+TEST(ObjectRequestsTest, List) {
+  ListObjectsRequest request("my-bucket");
+  EXPECT_EQ("my-bucket", request.bucket_name());
+  request.set_multiple_options(UserProject("my-project"), Prefix("foo/"));
+
+  std::ostringstream os;
+  os << request;
+  std::string actual = os.str();
+  EXPECT_THAT(actual, HasSubstr("my-bucket"));
+  EXPECT_THAT(actual, HasSubstr("userProject=my-project"));
+  EXPECT_THAT(actual, HasSubstr("prefix=foo/"));
+}
+
+TEST(ObjectRequestsTest, ParseListResponse) {
+  std::string object1 = R"""({
+      "bucket": "foo-bar",
+      "etag": "XYZ=",
+      "id": "baz",
+      "kind": "storage#object",
+      "generation": 1,
+      "location": "US",
+      "metadata": {
+        "foo": "bar",
+        "baz": "qux"
+      },
+      "metageneration": "4",
+      "name": "foo-bar-baz",
+      "projectNumber": "123456789",
+      "selfLink": "https://www.googleapis.com/storage/v1/b/foo-bar/baz/1",
+      "storageClass": "STANDARD",
+      "timeCreated": "2018-05-19T19:31:14Z",
+      "updated": "2018-05-19T19:31:24Z"
+})""";
+  std::string object2 = R"""({
+      "bucket": "foo-bar",
+      "etag": "XYZ=",
+      "id": "qux",
+      "kind": "storage#object",
+      "generation": "7",
+      "location": "US",
+      "metadata": {
+        "lbl1": "bar",
+        "lbl2": "qux"
+      },
+      "metageneration": "4",
+      "name": "qux",
+      "projectNumber": "123456789",
+      "selfLink": "https://www.googleapis.com/storage/v1/b/foo-bar/qux/7",
+      "storageClass": "STANDARD",
+      "timeCreated": "2018-05-19T19:31:14Z",
+      "updated": "2018-05-19T19:31:24Z"
+})""";
+  std::string text = R"""({
+      "kind": "storage#buckets",
+      "nextPageToken": "some-token-42",
+      "items":
+)""";
+  text += "[" + object1 + "," + object2 + "]}";
+
+  auto o1 = ObjectMetadata::ParseFromString(object1);
+  auto o2 = ObjectMetadata::ParseFromString(object2);
+
+  auto actual =
+      ListObjectsResponse::FromHttpResponse(HttpResponse{200, text, {}});
+  EXPECT_EQ("some-token-42", actual.next_page_token);
+  EXPECT_THAT(actual.items, ::testing::ElementsAre(o1, o2));
+}
+
+TEST(ObjectRequestsTest, Get) {
+  GetObjectMetadataRequest request("my-bucket", "my-object");
+  request.set_multiple_options(Generation(1), IfMetaGenerationMatch(3));
+  std::ostringstream os;
+  os << request;
+  auto str = os.str();
+  EXPECT_THAT(str, HasSubstr("my-bucket"));
+  EXPECT_THAT(str, HasSubstr("my-object"));
+  EXPECT_THAT(str, HasSubstr("generation=1"));
+  EXPECT_THAT(str, HasSubstr("ifMetagenerationMatch=3"));
+}
+
+TEST(ObjectRequestsTest, InsertObjectMedia) {
+  InsertObjectMediaRequest request("my-bucket", "my-object", "object contents");
+  request.set_multiple_options(
+      IfGenerationMatch(0), Projection("full"), ContentEncoding("media"),
+      KmsKeyName("random-key"), PredefinedAcl("authenticatedRead"));
+  std::ostringstream os;
+  os << request;
+  auto str = os.str();
+  EXPECT_THAT(str, HasSubstr("InsertObjectMediaRequest"));
+  EXPECT_THAT(str, HasSubstr("my-bucket"));
+  EXPECT_THAT(str, HasSubstr("my-object"));
+  EXPECT_THAT(str, HasSubstr("ifGenerationMatch=0"));
+  EXPECT_THAT(str, HasSubstr("projection=full"));
+  EXPECT_THAT(str, HasSubstr("kmsKeyName=random-key"));
+  EXPECT_THAT(str, HasSubstr("contentEncoding=media"));
+  EXPECT_THAT(str, HasSubstr("predefinedAcl=authenticatedRead"));
+}
+
+TEST(ObjectRequestsTest, InsertObjectStreaming) {
+  InsertObjectStreamingRequest request("my-bucket", "my-object");
+  request.set_multiple_options(
+      IfGenerationMatch(0), Projection("full"), ContentEncoding("media"),
+      KmsKeyName("random-key"), PredefinedAcl("authenticatedRead"));
+  std::ostringstream os;
+  os << request;
+  auto str = os.str();
+  EXPECT_THAT(str, HasSubstr("InsertObjectStreamingRequest"));
+  EXPECT_THAT(str, HasSubstr("my-bucket"));
+  EXPECT_THAT(str, HasSubstr("my-object"));
+  EXPECT_THAT(str, HasSubstr("ifGenerationMatch=0"));
+  EXPECT_THAT(str, HasSubstr("projection=full"));
+  EXPECT_THAT(str, HasSubstr("kmsKeyName=random-key"));
+  EXPECT_THAT(str, HasSubstr("contentEncoding=media"));
+  EXPECT_THAT(str, HasSubstr("predefinedAcl=authenticatedRead"));
+}
+
+HttpResponse CreateRangeRequestResponse(
+    char const* content_range_header_value) {
+  HttpResponse response;
+  response.headers.emplace(std::string("content-range"),
+                           std::string(content_range_header_value));
+  response.payload = "some payload";
+  return response;
+}
+
+TEST(ObjectRequestsTest, ReadObjectRange) {
+  ReadObjectRangeRequest request("my-bucket", "my-object", 0, 1024);
+
+  EXPECT_EQ("my-bucket", request.bucket_name());
+  EXPECT_EQ("my-object", request.object_name());
+  EXPECT_EQ(0, request.begin());
+  EXPECT_EQ(1024, request.end());
+
+  request.set_option(storage::UserProject("my-project"));
+  request.set_multiple_options(storage::IfGenerationMatch(7),
+                               storage::UserProject("my-project"));
+
+  std::ostringstream os;
+  os << request;
+  std::string actual = os.str();
+  EXPECT_THAT(actual, HasSubstr("my-bucket"));
+  EXPECT_THAT(actual, HasSubstr("my-object"));
+  EXPECT_THAT(actual, HasSubstr("begin=0"));
+  EXPECT_THAT(actual, HasSubstr("end=1024"));
+  EXPECT_THAT(actual, HasSubstr("ifGenerationMatch=7"));
+  EXPECT_THAT(actual, HasSubstr("my-project"));
+}
+
+TEST(ObjectRequestsTest, RangeResponseParse) {
+  auto actual = ReadObjectRangeResponse::FromHttpResponse(
+      CreateRangeRequestResponse("bytes 100-200/20000"));
+  EXPECT_EQ(100, actual.first_byte);
+  EXPECT_EQ(200, actual.last_byte);
+  EXPECT_EQ(20000, actual.object_size);
+  EXPECT_EQ("some payload", actual.contents);
+}
+
+TEST(ObjectRequestsTest, RangeResponseParseStar) {
+  auto actual = ReadObjectRangeResponse::FromHttpResponse(
+      CreateRangeRequestResponse("bytes */20000"));
+  EXPECT_EQ(0, actual.first_byte);
+  EXPECT_EQ(0, actual.last_byte);
+  EXPECT_EQ(20000, actual.object_size);
+}
+
+TEST(ObjectRequestsTest, RangeResponseParseErrors) {
+#if GOOGLE_CLOUD_CPP_HAVE_EXCEPTIONS
+  EXPECT_THROW(ReadObjectRangeResponse::FromHttpResponse(
+                   CreateRangeRequestResponse("bits 100-200/20000")),
+               std::invalid_argument);
+  EXPECT_THROW(ReadObjectRangeResponse::FromHttpResponse(
+                   CreateRangeRequestResponse("100-200/20000")),
+               std::invalid_argument);
+  EXPECT_THROW(ReadObjectRangeResponse::FromHttpResponse(
+                   CreateRangeRequestResponse("bytes ")),
+               std::invalid_argument);
+  EXPECT_THROW(ReadObjectRangeResponse::FromHttpResponse(
+                   CreateRangeRequestResponse("bytes */")),
+               std::invalid_argument);
+  EXPECT_THROW(ReadObjectRangeResponse::FromHttpResponse(
+                   CreateRangeRequestResponse("bytes 100-200/")),
+               std::invalid_argument);
+  EXPECT_THROW(ReadObjectRangeResponse::FromHttpResponse(
+                   CreateRangeRequestResponse("bytes 100-/20000")),
+               std::invalid_argument);
+  EXPECT_THROW(ReadObjectRangeResponse::FromHttpResponse(
+                   CreateRangeRequestResponse("bytes -200/20000")),
+               std::invalid_argument);
+#else
+  EXPECT_DEATH_IF_SUPPORTED(
+      ReadObjectRangeResponse::FromHttpResponse(
+          CreateRangeRequestResponse("bits 100-200/20000")),
+      "exceptions are disabled");
+  EXPECT_DEATH_IF_SUPPORTED(ReadObjectRangeResponse::FromHttpResponse(
+                                CreateRangeRequestResponse("100-200/20000")),
+                            "exceptions are disabled");
+  EXPECT_DEATH_IF_SUPPORTED(ReadObjectRangeResponse::FromHttpResponse(
+                                CreateRangeRequestResponse("bytes ")),
+                            "exceptions are disabled");
+  EXPECT_DEATH_IF_SUPPORTED(ReadObjectRangeResponse::FromHttpResponse(
+                                CreateRangeRequestResponse("bytes */")),
+                            "exceptions are disabled");
+  EXPECT_DEATH_IF_SUPPORTED(ReadObjectRangeResponse::FromHttpResponse(
+                                CreateRangeRequestResponse("bytes 100-200/")),
+                            "exceptions are disabled");
+  EXPECT_DEATH_IF_SUPPORTED(ReadObjectRangeResponse::FromHttpResponse(
+                                CreateRangeRequestResponse("bytes 100-/20000")),
+                            "exceptions are disabled");
+  EXPECT_DEATH_IF_SUPPORTED(ReadObjectRangeResponse::FromHttpResponse(
+                                CreateRangeRequestResponse("bytes -200/20000")),
+                            "exceptions are disabled");
+#endif  // GOOGLE_CLOUD_CPP_HAVE_EXCEPTIONS
+}
+
+TEST(ObjectRequestsTest, Delete) {
+  DeleteObjectRequest request("my-bucket", "my-object");
+  request.set_multiple_options(IfMetaGenerationNotMatch(7),
+                               UserProject("my-project"));
+  std::ostringstream os;
+  os << request;
+  EXPECT_THAT(os.str(), HasSubstr("my-bucket"));
+  EXPECT_THAT(os.str(), HasSubstr("my-object"));
+  EXPECT_THAT(os.str(), HasSubstr("ifMetagenerationNotMatch=7"));
+  EXPECT_THAT(os.str(), HasSubstr("userProject=my-project"));
+}
+
+}  // namespace
+}  // namespace internal
+}  // namespace STORAGE_CLIENT_NS
+}  // namespace storage
+}  // namespace cloud
+}  // namespace google
