+++ google/cloud/storage/storage_client.bzl
@@ -22,26 +22,22 @@ storage_client_HDRS = [
     "internal/curl_client.h",
     "internal/curl_streambuf.h",
     "internal/default_object_acl_requests.h",
-    "internal/delete_object_request.h",
     "internal/empty_response.h",
     "internal/format_rfc3339.h",
     "internal/generic_object_request.h",
     "internal/generic_request.h",
-    "internal/get_object_metadata_request.h",
     "internal/google_application_default_credentials_file.h",
     "internal/http_response.h",
-    "internal/insert_object_media_request.h",
-    "internal/list_objects_request.h",
     "internal/logging_client.h",
     "internal/metadata_parser.h",
     "internal/nljson.h",
     "internal/openssl_util.h",
     "internal/object_acl_requests.h",
+    "internal/object_requests.h",
     "internal/object_streambuf.h",
     "internal/parse_rfc3339.h",
     "internal/patch_builder.h",
     "internal/raw_client.h",
-    "internal/read_object_range_request.h",
     "internal/retry_client.h",
     "internal/service_account_credentials.h",
     "lifecycle_rule.h",
@@ -77,19 +73,15 @@ storage_client_SRCS = [
     "internal/curl_client.cc",
     "internal/curl_streambuf.cc",
     "internal/default_object_acl_requests.cc",
-    "internal/delete_object_request.cc",
     "internal/empty_response.cc",
     "internal/format_rfc3339.cc",
-    "internal/get_object_metadata_request.cc",
     "internal/google_application_default_credentials_file.cc",
-    "internal/insert_object_media_request.cc",
-    "internal/list_objects_request.cc",
     "internal/logging_client.cc",
     "internal/metadata_parser.cc",
     "internal/object_acl_requests.cc",
+    "internal/object_requests.cc",
     "internal/object_streambuf.cc",
     "internal/parse_rfc3339.cc",
-    "internal/read_object_range_request.cc",
     "internal/retry_client.cc",
     "lifecycle_rule.cc",
     "list_buckets_reader.cc",
