+++ google/cloud/storage/storage_client_unit_tests.bzl
@@ -15,20 +15,16 @@ storage_client_unit_tests = [
     "internal/bucket_acl_requests_test.cc",
     "internal/bucket_requests_test.cc",
     "internal/default_object_acl_requests_test.cc",
-    "internal/delete_object_request_test.cc",
     "internal/format_rfc3339_test.cc",
-    "internal/get_object_metadata_request_test.cc",
     "internal/google_application_default_credentials_file_test.cc",
-    "internal/insert_object_media_request_test.cc",
-    "internal/list_objects_request_test.cc",
     "internal/logging_client_test.cc",
     "internal/metadata_parser_test.cc",
     "internal/nljson_test.cc",
     "internal/object_acl_requests_test.cc",
+    "internal/object_requests_test.cc",
     "internal/parse_rfc3339_test.cc",
     "internal/patch_builder_test.cc",
     "internal/retry_client_test.cc",
-    "internal/read_object_range_request_test.cc",
     "internal/service_account_credentials_test.cc",
     "lifecycle_rule_test.cc",
     "list_buckets_reader_test.cc",
